# openapi.json
{"openapi":"3.1.0","info":{"title":"LlamaIndex Chat API","description":"Chat interface with LlamaIndex + MinerU for multi-source document retrieval","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health Check","description":"Health check endpoint.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthCheckResponse"}}}}}}},"/":{"get":{"summary":"Root","description":"Root endpoint.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/chat/sessions":{"post":{"tags":["chat"],"summary":"Create Session","description":"Create a new chat session.\n\nArgs:\n    request: Session creation request\n\nReturns:\n    Session information","operationId":"create_session_api_chat_sessions_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSessionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Create Session Api Chat Sessions Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["chat"],"summary":"List Sessions","description":"List all chat sessions.\n\nArgs:\n    limit: Maximum number of sessions to return\n    offset: Number of sessions to skip\n\nReturns:\n    List of session summaries","operationId":"list_sessions_api_chat_sessions_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chat/sessions/{session_id}":{"get":{"tags":["chat"],"summary":"Get Session","description":"Get session details.\n\nArgs:\n    session_id: Session ID\n\nReturns:\n    Session information with message history","operationId":"get_session_api_chat_sessions__session_id__get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["chat"],"summary":"Delete Session","description":"Delete a chat session.\n\nArgs:\n    session_id: Session ID\n\nReturns:\n    Deletion confirmation","operationId":"delete_session_api_chat_sessions__session_id__delete","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chat/message":{"post":{"tags":["chat"],"summary":"Send Message","description":"Send a message and get response.\n\nArgs:\n    request: Message request\n\nReturns:\n    Chat response with sources","operationId":"send_message_api_chat_message_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendMessageRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chat/sessions/{session_id}/history":{"get":{"tags":["chat"],"summary":"Get Conversation History","description":"Get conversation history for a session.\n\nArgs:\n    session_id: Session ID\n    max_messages: Maximum number of recent messages\n\nReturns:\n    List of messages","operationId":"get_conversation_history_api_chat_sessions__session_id__history_get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}},{"name":"max_messages","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","maximum":100,"minimum":1},{"type":"null"}],"title":"Max Messages"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/index/stats":{"get":{"tags":["index"],"summary":"Get Index Stats","description":"Get index statistics.\n\nReturns:\n    Index statistics including document count, retrieval modes, etc.","operationId":"get_index_stats_api_index_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndexStatsResponse"}}}}}}},"/api/index/build":{"post":{"tags":["index"],"summary":"Build Index","description":"Build or rebuild the index.\n\nNote: This endpoint is for rebuilding the entire index.\nFor adding documents, use the source routes.\n\nReturns:\n    Build status","operationId":"build_index_api_index_build_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildIndexResponse"}}}}}}},"/api/index/clear":{"delete":{"tags":["index"],"summary":"Clear Index","description":"Clear all documents from the index.\n\nWarning: This operation cannot be undone.\n\nReturns:\n    Confirmation message","operationId":"clear_index_api_index_clear_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/index/status":{"get":{"tags":["index"],"summary":"Get Index Status","description":"Get index status and health.\n\nReturns:\n    Index status information","operationId":"get_index_status_api_index_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/sources/upload":{"post":{"tags":["sources"],"summary":"Upload File","description":"Upload and index a file.\n\nSupports: PDF, DOCX, XLSX, PPTX\n\nArgs:\n    file: Uploaded file\n\nReturns:\n    Upload status and document count","operationId":"upload_file_api_sources_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_api_sources_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadFileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sources/jira/test":{"post":{"tags":["sources"],"summary":"Test Jira Connection","description":"Test Jira connection.\n\nArgs:\n    request: Jira connection details\n\nReturns:\n    Connection test result","operationId":"test_jira_connection_api_sources_jira_test_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JiraConnectionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConnectionTestResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sources/jira/sync":{"post":{"tags":["sources"],"summary":"Sync Jira","description":"Sync Jira issues and add to index.\n\nArgs:\n    request: Jira connection and query details\n\nReturns:\n    Sync status","operationId":"sync_jira_api_sources_jira_sync_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JiraConnectionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SyncSourceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sources/confluence/test":{"post":{"tags":["sources"],"summary":"Test Confluence Connection","description":"Test Confluence connection.\n\nArgs:\n    request: Confluence connection details\n\nReturns:\n    Connection test result","operationId":"test_confluence_connection_api_sources_confluence_test_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfluenceConnectionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConnectionTestResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sources/confluence/sync":{"post":{"tags":["sources"],"summary":"Sync Confluence","description":"Sync Confluence pages and add to index.\n\nArgs:\n    request: Confluence connection and query details\n\nReturns:\n    Sync status","operationId":"sync_confluence_api_sources_confluence_sync_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfluenceConnectionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SyncSourceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sources/list":{"get":{"tags":["sources"],"summary":"List Sources","description":"List all configured data sources.\n\nReturns:\n    List of data sources","operationId":"list_sources_api_sources_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/analysis/issue/{issue_key}":{"post":{"tags":["analysis"],"summary":"Analyze Issue","description":"Analyze a Jira issue.\n\nArgs:\n    issue_key: Issue key (e.g., \"PROJ-123\")\n    request: Analysis request parameters (optional)\n\nReturns:\n    IssueAnalysisResult with analysis and sources","operationId":"analyze_issue_api_analysis_issue__issue_key__post","parameters":[{"name":"issue_key","in":"path","required":true,"schema":{"type":"string","title":"Issue Key"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalysisRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IssueAnalysisResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["analysis"],"summary":"Get Saved Analysis","description":"Get saved analysis for an issue.\n\nArgs:\n    issue_key: Issue key\n\nReturns:\n    Saved analysis data","operationId":"get_saved_analysis_api_analysis_issue__issue_key__get","parameters":[{"name":"issue_key","in":"path","required":true,"schema":{"type":"string","title":"Issue Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Saved Analysis Api Analysis Issue  Issue Key  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analysis/issues":{"get":{"tags":["analysis"],"summary":"List Analyzed Issues","description":"List all analyzed issues.\n\nReturns:\n    List of issue metadata","operationId":"list_analyzed_issues_api_analysis_issues_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Response List Analyzed Issues Api Analysis Issues Get"}}}}}}},"/api/reports/daily":{"post":{"tags":["reports"],"summary":"Generate Daily Report","description":"Generate a daily report.\n\nArgs:\n    request: Report generation request\n\nReturns:\n    ReportResponse with report_id and initial status","operationId":"generate_daily_report_api_reports_daily_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reports/daily/{report_id}":{"get":{"tags":["reports"],"summary":"Get Daily Report","description":"Get a daily report by ID.\n\nArgs:\n    report_id: Report identifier\n\nReturns:\n    ReportResponse with current status","operationId":"get_daily_report_api_reports_daily__report_id__get","parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"string","title":"Report Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reports/daily/saved/{date}":{"get":{"tags":["reports"],"summary":"Get Saved Report","description":"Get a saved daily report from knowledge base.\n\nArgs:\n    date: Date in YYYY-MM-DD format\n\nReturns:\n    Saved report data","operationId":"get_saved_report_api_reports_daily_saved__date__get","parameters":[{"name":"date","in":"path","required":true,"schema":{"type":"string","title":"Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reports/daily/list":{"get":{"tags":["reports"],"summary":"List Saved Reports","description":"List all saved daily reports.\n\nReturns:\n    List of report dates and metadata","operationId":"list_saved_reports_api_reports_daily_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AnalysisRequest":{"properties":{"issue_key":{"type":"string","title":"Issue Key","description":"Issue key to analyze"},"depth":{"type":"string","title":"Depth","description":"Analysis depth (quick or deep)","default":"deep"},"include_related":{"type":"boolean","title":"Include Related","description":"Include related issues","default":true},"save_to_kb":{"type":"boolean","title":"Save To Kb","description":"Save to knowledge base","default":true}},"type":"object","required":["issue_key"],"title":"AnalysisRequest","description":"Request for issue analysis."},"AnalysisSource":{"properties":{"source_id":{"type":"string","title":"Source Id","description":"Source document ID"},"title":{"type":"string","title":"Title","description":"Document title"},"snippet":{"type":"string","title":"Snippet","description":"Relevant snippet"},"score":{"type":"number","title":"Score","description":"Relevance score"},"source_type":{"type":"string","title":"Source Type","description":"Source type (pdf, jira, confluence)"}},"type":"object","required":["source_id","title","snippet","score","source_type"],"title":"AnalysisSource","description":"Source document used in analysis."},"Body_upload_file_api_sources_upload_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_file_api_sources_upload_post"},"BuildIndexResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"message":{"type":"string","title":"Message"},"documents_added":{"type":"integer","title":"Documents Added"},"collection_name":{"type":"string","title":"Collection Name"}},"type":"object","required":["success","message","documents_added","collection_name"],"title":"BuildIndexResponse","description":"Response after building index."},"ChatResponse":{"properties":{"message_id":{"type":"string","title":"Message Id","description":"Message identifier"},"content":{"type":"string","title":"Content","description":"Assistant response"},"sources":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Sources","description":"Source citations"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp","description":"Response timestamp"},"metadata":{"additionalProperties":true,"type":"object","title":"Metadata","description":"Response metadata"}},"type":"object","required":["message_id","content"],"title":"ChatResponse","description":"Response from chat engine."},"ConfluenceConnectionRequest":{"properties":{"base_url":{"type":"string","title":"Base Url","description":"Confluence base URL"},"email":{"type":"string","title":"Email","description":"User email"},"api_token":{"type":"string","title":"Api Token","description":"API token"},"space_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Space Key","description":"Space key (optional)"},"page_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Page Id","description":"Page ID (optional)"},"max_pages":{"type":"integer","title":"Max Pages","description":"Maximum pages","default":100}},"type":"object","required":["base_url","email","api_token"],"title":"ConfluenceConnectionRequest","description":"Request to connect to Confluence."},"ConnectionTestResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"message":{"type":"string","title":"Message"},"user":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User"}},"type":"object","required":["success","message"],"title":"ConnectionTestResponse","description":"Response from connection test."},"CreateSessionRequest":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name","description":"Session name"},"metadata":{"additionalProperties":true,"type":"object","title":"Metadata","description":"Session metadata"}},"type":"object","title":"CreateSessionRequest","description":"Request to create a new session."},"FullReport":{"properties":{"date":{"type":"string","title":"Date"},"quick_report":{"$ref":"#/components/schemas/QuickReport"},"summary":{"type":"string","title":"Summary"},"key_updates":{"items":{"type":"string"},"type":"array","title":"Key Updates"},"recommendations":{"items":{"type":"string"},"type":"array","title":"Recommendations"},"generated_at":{"type":"string","title":"Generated At"}},"type":"object","required":["date","quick_report","summary","key_updates","recommendations"],"title":"FullReport","description":"Full daily report with LLM analysis."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthCheckResponse":{"properties":{"status":{"type":"string","title":"Status","description":"Service status"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"version":{"type":"string","title":"Version","default":"0.1.0"},"components":{"additionalProperties":{"type":"string"},"type":"object","title":"Components"}},"type":"object","required":["status"],"title":"HealthCheckResponse","description":"Health check response."},"IndexStatsResponse":{"properties":{"collection_name":{"type":"string","title":"Collection Name"},"total_documents":{"type":"integer","title":"Total Documents"},"vector_store":{"additionalProperties":true,"type":"object","title":"Vector Store"},"bm25_enabled":{"type":"boolean","title":"Bm25 Enabled"},"hybrid_enabled":{"type":"boolean","title":"Hybrid Enabled"},"retrieval_modes":{"items":{"type":"string"},"type":"array","title":"Retrieval Modes"}},"type":"object","required":["collection_name","total_documents","vector_store","bm25_enabled","hybrid_enabled","retrieval_modes"],"title":"IndexStatsResponse","description":"Response with index statistics."},"IssueAnalysisResult":{"properties":{"issue_key":{"type":"string","title":"Issue Key","description":"Issue key"},"issue_summary":{"type":"string","title":"Issue Summary","description":"Issue summary"},"issue_description":{"type":"string","title":"Issue Description","description":"Issue description"},"analysis":{"type":"string","title":"Analysis","description":"Generated analysis (Markdown format)"},"sources":{"items":{"$ref":"#/components/schemas/AnalysisSource"},"type":"array","title":"Sources","description":"Source documents"},"related_issues":{"items":{"$ref":"#/components/schemas/RelatedIssue"},"type":"array","title":"Related Issues","description":"Related issues"},"timestamp":{"type":"string","title":"Timestamp","description":"Analysis timestamp"},"depth":{"type":"string","title":"Depth","description":"Analysis depth (quick or deep)","default":"deep"}},"type":"object","required":["issue_key","issue_summary","issue_description","analysis"],"title":"IssueAnalysisResult","description":"Result of issue analysis."},"IssueStats":{"properties":{"key":{"type":"string","title":"Key"},"summary":{"type":"string","title":"Summary"},"status":{"type":"string","title":"Status"},"priority":{"type":"string","title":"Priority"},"assignee":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Assignee"},"updated":{"type":"string","title":"Updated"}},"type":"object","required":["key","summary","status","priority","updated"],"title":"IssueStats","description":"Statistics for a single issue."},"JiraConnectionRequest":{"properties":{"base_url":{"type":"string","title":"Base Url","description":"Jira base URL"},"email":{"type":"string","title":"Email","description":"User email"},"api_token":{"type":"string","title":"Api Token","description":"API token"},"jql":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Jql","description":"JQL query (optional)"},"project_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Project Key","description":"Project key (optional)"},"max_results":{"type":"integer","title":"Max Results","description":"Maximum results","default":100}},"type":"object","required":["base_url","email","api_token"],"title":"JiraConnectionRequest","description":"Request to connect to Jira."},"QuickReport":{"properties":{"date":{"type":"string","title":"Date"},"issues":{"items":{"$ref":"#/components/schemas/IssueStats"},"type":"array","title":"Issues"},"stats":{"$ref":"#/components/schemas/QuickReportStats"},"generated_at":{"type":"string","title":"Generated At"}},"type":"object","required":["date","issues","stats"],"title":"QuickReport","description":"Quick daily report (no LLM analysis)."},"QuickReportStats":{"properties":{"total_issues":{"type":"integer","title":"Total Issues"},"by_status":{"additionalProperties":{"type":"integer"},"type":"object","title":"By Status"},"by_priority":{"additionalProperties":{"type":"integer"},"type":"object","title":"By Priority"},"by_assignee":{"additionalProperties":{"type":"integer"},"type":"object","title":"By Assignee"}},"type":"object","required":["total_issues"],"title":"QuickReportStats","description":"Quick statistics for daily report."},"RelatedIssue":{"properties":{"issue_key":{"type":"string","title":"Issue Key","description":"Issue key (e.g., PROJ-123)"},"summary":{"type":"string","title":"Summary","description":"Issue summary"},"status":{"type":"string","title":"Status","description":"Issue status"},"link_type":{"type":"string","title":"Link Type","description":"Link type (e.g., blocks, relates to)"},"url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Url","description":"Issue URL"}},"type":"object","required":["issue_key","summary","status","link_type"],"title":"RelatedIssue","description":"Related Jira issue."},"ReportRequest":{"properties":{"date":{"type":"string","title":"Date"},"mode":{"type":"string","title":"Mode","default":"quick"}},"type":"object","required":["date"],"title":"ReportRequest","description":"Request to generate a daily report."},"ReportResponse":{"properties":{"report_id":{"type":"string","title":"Report Id"},"status":{"type":"string","title":"Status"},"quick_report":{"anyOf":[{"$ref":"#/components/schemas/QuickReport"},{"type":"null"}]},"full_report":{"anyOf":[{"$ref":"#/components/schemas/FullReport"},{"type":"null"}]},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["report_id","status"],"title":"ReportResponse","description":"Response for report generation."},"SendMessageRequest":{"properties":{"session_id":{"type":"string","title":"Session Id","description":"Session ID"},"message":{"type":"string","title":"Message","description":"User message"},"source_filters":{"additionalProperties":true,"type":"object","title":"Source Filters","description":"Filters for source retrieval"},"retrieval_mode":{"type":"string","title":"Retrieval Mode","description":"Retrieval mode: 'hybrid', 'vector', or 'bm25'","default":"hybrid"},"similarity_top_k":{"type":"integer","title":"Similarity Top K","description":"Number of sources to retrieve","default":5}},"type":"object","required":["session_id","message"],"title":"SendMessageRequest","description":"Request to send a message."},"SessionListResponse":{"properties":{"sessions":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Sessions","description":"List of session summaries"},"total":{"type":"integer","title":"Total","description":"Total number of sessions"}},"type":"object","required":["sessions","total"],"title":"SessionListResponse","description":"Response with list of sessions."},"SyncSourceResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"message":{"type":"string","title":"Message"},"documents_synced":{"type":"integer","title":"Documents Synced"},"source_id":{"type":"string","title":"Source Id"}},"type":"object","required":["success","message","documents_synced","source_id"],"title":"SyncSourceResponse","description":"Response after syncing source."},"UploadFileResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"message":{"type":"string","title":"Message"},"file_path":{"type":"string","title":"File Path"},"documents_created":{"type":"integer","title":"Documents Created"}},"type":"object","required":["success","message","file_path","documents_created"],"title":"UploadFileResponse","description":"Response after file upload."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}